Split data and enum classes into separate files
