+++ AlgorithmType.java
@@ -1,0 +1,5 @@
+package com.yumsg.core.enums;
+
+public enum AlgorithmType {
+    ASYMMETRIC_KEM, SYMMETRIC, DIGITAL_SIGNATURE
+}
