+++ AppScreen.java
@@ -1,0 +1,5 @@
+package com.yumsg.core.enums;
+
+public enum AppScreen {
+    MODE_SELECTION, SERVER_CONNECTION, AUTHENTICATION, CHAT_LIST, CHAT_ACTIVE
+}
