+++ AppState.java
@@ -1,0 +1,6 @@
+package com.yumsg.core.enums;
+
+public enum AppState {
+    INITIALIZING, MODE_SELECTION, SERVER_CONNECTION, AUTHENTICATION,
+    CHAT_LIST, CHAT_ACTIVE, MODE_SELECTED, AUTHENTICATED, RESTORED
+}
