+++ Contact.java
@@ -1,0 +1,16 @@
+package com.yumsg.core.data;
+
+public class Contact {
+    private String id;
+    private String name;
+    private String publicKey;
+
+    public Contact() {}
+
+    public String getId() { return id; }
+    public void setId(String id) { this.id = id; }
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+    public String getPublicKey() { return publicKey; }
+    public void setPublicKey(String publicKey) { this.publicKey = publicKey; }
+}
