+++ Permission.java
@@ -1,0 +1,5 @@
+package com.yumsg.core.enums;
+
+public enum Permission {
+    CAMERA, MICROPHONE, STORAGE, LOCATION, NOTIFICATIONS, CONTACTS
+}
