+++ SoundType.java
@@ -1,0 +1,5 @@
+package com.yumsg.core.enums;
+
+public enum SoundType {
+    MESSAGE_RECEIVED, MESSAGE_SENT, NOTIFICATION, ERROR, SUCCESS
+}
