+++ ThemeMode.java
@@ -1,0 +1,5 @@
+package com.yumsg.core.enums;
+
+public enum ThemeMode {
+    LIGHT, DARK, SYSTEM
+}
