+++ UserProfile.java
@@ -1,0 +1,19 @@
+package com.yumsg.core.data;
+
+public class UserProfile {
+    private String id;
+    private String username;
+    private String email;
+    private String displayName;
+
+    public UserProfile() {}
+
+    public String getId() { return id; }
+    public void setId(String id) { this.id = id; }
+    public String getUsername() { return username; }
+    public void setUsername(String username) { this.username = username; }
+    public String getEmail() { return email; }
+    public void setEmail(String email) { this.email = email; }
+    public String getDisplayName() { return displayName; }
+    public void setDisplayName(String displayName) { this.displayName = displayName; }
+}
