+++ UserRegistrationInfo.java
@@ -1,0 +1,23 @@
+package com.yumsg.core.data;
+
+public class UserRegistrationInfo {
+    private String username;
+    private String email;
+    private String password;
+    private String displayName;
+    private String organizationCode;
+
+    public UserRegistrationInfo(String username, String email, String password, String displayName) {
+        this.username = username;
+        this.email = email;
+        this.password = password;
+        this.displayName = displayName;
+    }
+
+    public String getUsername() { return username; }
+    public String getEmail() { return email; }
+    public String getPassword() { return password; }
+    public String getDisplayName() { return displayName; }
+    public String getOrganizationCode() { return organizationCode; }
+    public void setOrganizationCode(String organizationCode) { this.organizationCode = organizationCode; }
+}
