+++ UserStatus.java
@@ -1,0 +1,5 @@
+package com.yumsg.core.enums;
+
+public enum UserStatus {
+    ONLINE, OFFLINE, AWAY, BUSY
+}
